refactor: :recycle: move `/scripts` to `/src/cli`
