+++ src/cli/balance.ts
