+++ src/cli/delete.ts
