+++ src/cli/test.ts
