+++ src/index.ts
@@ -1,14 +1,14 @@
 import { Command } from 'commander'
 
-import balance from './balance.js'
-import del from './delete.js'
-import exp from './export.js'
-import imp from './import.js'
-import insert from './insert.js'
-import tags from './tags.js'
-import template from './template.js'
-import test from './test.js'
-import txs from './txs.js'
+import balance from './cli/balance.js'
+import del from './cli/delete.js'
+import exp from './cli/export.js'
+import imp from './cli/import.js'
+import insert from './cli/insert.js'
+import tags from './cli/tags.js'
+import template from './cli/template.js'
+import test from './cli/test.js'
+import txs from './cli/txs.js'
 
 const program = new Command()
