+++ tsconfig.json
@@ -106,7 +106,7 @@
 		"skipLibCheck": true /* Skip type checking all .d.ts files. */,
 		"paths": {
 			"#/*": ["./src/*"],
-			"#scripts/*": ["./scripts/*"]
+			"#scripts/*": ["src/scripts/*"]
 		}
 	}
 }
